refactor: rename and move function logic into helper
